rework color ctlseqs
output direct color sequences that use colon ':' as the parameter
substring separator -- this should be in line with ISO-8613-6 (although
I don't have access to that), based on Thomas Dickey's documentation.
this is potentially incompatible with some terminals/MUD clients that
did support the "incorrect" semicolon separator, but hopefully that's
not the case in 2020.

in addition, use a statically allocated buffer for the color ctlseq
output -- we never need more than one at a time anyway, so it makes no
sense to use asprintf here.