revert back to CSI 38 ; Pr ; Pg ; Pb for 24bit color
previous commit turned out to be too optimistic; the form
CSI 38 : Pi : Pr : Pb : Pg does not seem to work with eg. st 0.8.4 or
alacritty 0.3.3. XTerm(351) and tmux openbsd-6.7 do understand it, but
in the interest if compatibility, let's stick with the form that most
terminals implement, even if incorrect by standard.